+++ packages/code-du-travail-frontend/cypress/integration/light/widgets/dossier-licenciement.spec.ts
@@ -1,8 +1,10 @@
+import "cypress-iframe";
+
 describe("Widget - Dossier Licenciement", () => {
   it("should display the widget", () => {
     cy.visit("https://socialgouv.github.io/cdtn-admin/procedure-licenciement");
-    cy.contains("Comprendre sa procédure de licenciement", {
-      timeout: 10000,
-    });
+    cy.iframe()
+      .contains("Comprendre sa procédure de licenciement")
+      .should("be.visible");
   });
 });
