+++ packages/code-du-travail-frontend/cypress/integration/light/widgets/indemnite-licenciement.spec.ts
@@ -1,8 +1,10 @@
+import "cypress-iframe";
+
 describe("Widget - Indemnité de licenciement", () => {
   it("should display the widget", () => {
     cy.visit("https://socialgouv.github.io/cdtn-admin/indemnite-licenciement");
-    cy.contains("Calculer l'indemnité de licenciement", {
-      timeout: 10000,
-    });
+    cy.iframe()
+      .contains("Calculer l'indemnité de licenciement")
+      .should("be.visible");
   });
 });
