+++ packages/code-du-travail-frontend/cypress/integration/light/widgets/indemnite-precarite.spec.ts
@@ -1,8 +1,10 @@
+import "cypress-iframe";
+
 describe("Widget - Indemnité de Precarite", () => {
   it("should display the widget", () => {
     cy.visit("https://socialgouv.github.io/cdtn-admin/indemnite-precarite");
-    cy.contains("Calculer l'indemnité de précarité", {
-      timeout: 10000,
-    });
+    cy.iframe()
+      .contains("Calculer l'indemnité de précarité")
+      .should("be.visible");
   });
 });
