+++ packages/code-du-travail-frontend/cypress/integration/light/widgets/modeles-courrier.spec.ts
@@ -1,8 +1,10 @@
+import "cypress-iframe";
+
 describe("Widget - Modèles de courrier", () => {
   it("should display the widget", () => {
     cy.visit("https://socialgouv.github.io/cdtn-admin/modeles");
-    cy.contains("LUTTE CONTRE LE HARCELEMENT SEXUEL", {
-      timeout: 10000,
-    });
+    cy.iframe()
+      .contains("LUTTE CONTRE LE HARCELEMENT SEXUEL")
+      .should("be.visible");
   });
 });
