+++ packages/code-du-travail-frontend/cypress/integration/light/widgets/preavis-demission.spec.ts
@@ -1,8 +1,10 @@
+import "cypress-iframe";
+
 describe("Widget - Préavis de démission", () => {
   it("should display the widget", () => {
     cy.visit("https://socialgouv.github.io/cdtn-admin/preavis-demission");
-    cy.contains("Calculer le préavis de démission", {
-      timeout: 10000,
-    });
+    cy.iframe()
+      .contains("Calculer le préavis de démission")
+      .should("be.visible");
   });
 });
