+++ packages/code-du-travail-frontend/cypress/integration/light/widgets/preavis-licenciement.spec.ts
@@ -1,8 +1,10 @@
+import "cypress-iframe";
+
 describe("Widget - Préavis de licenciement", () => {
   it("should display the widget", () => {
     cy.visit("https://socialgouv.github.io/cdtn-admin/preavis-licenciement");
-    cy.contains("Calculer le préavis de licenciement", {
-      timeout: 10000,
-    });
+    cy.iframe()
+      .contains("Calculer le préavis de licenciement")
+      .should("be.visible");
   });
 });
