+++ packages/code-du-travail-frontend/cypress/integration/light/widgets/preavis-retraite.spec.ts
@@ -1,8 +1,10 @@
+import "cypress-iframe";
+
 describe("Widget - Préavis de retraite", () => {
   it("should display the widget", () => {
     cy.visit("https://socialgouv.github.io/cdtn-admin/preavis-retraite");
-    cy.contains("Calculer le préavis de départ à la retraite", {
-      timeout: 10000,
-    });
+    cy.iframe()
+      .contains("Calculer le préavis de départ à la retraite")
+      .should("be.visible");
   });
 });
